Add doc and rename to windowsevent
